[CST-5306] Improvements on ResearcherProfileRestRepository

+++ dspace-api/src/main/java/org/dspace/app/profile/ResearcherProfileServiceImpl.java
@@ -9,7 +9,6 @@ package org.dspace.app.profile;
 
 import static java.util.Optional.empty;
 import static java.util.Optional.ofNullable;
-import static org.apache.commons.lang3.ArrayUtils.contains;
 import static org.dspace.content.authority.Choices.CF_ACCEPTED;
 import static org.dspace.core.Constants.READ;
 import static org.dspace.eperson.Group.ANONYMOUS;
@@ -24,7 +23,7 @@ import java.util.UUID;
 
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
-import org.dspace.app.exception.ResourceConflictException;
+import org.dspace.app.exception.ResourceAlreadyExistsException;
 import org.dspace.app.profile.service.ResearcherProfileService;
 import org.dspace.authorize.AuthorizeException;
 import org.dspace.authorize.service.AuthorizeService;
@@ -105,8 +104,7 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
 
         Item profileItem = findResearcherProfileItemById(context, ePerson.getID());
         if (profileItem != null) {
-            ResearcherProfile profile = new ResearcherProfile(profileItem);
-            throw new ResourceConflictException("A profile is already linked to the provided User", profile);
+            throw new ResourceAlreadyExistsException("A profile is already linked to the provided User");
         }
 
         Collection collection = findProfileCollection(context)
@@ -136,7 +134,7 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
         if (isHardDeleteEnabled()) {
             deleteItem(context, profileItem);
         } else {
-            removeDspaceObjectOwnerMetadata(context, profileItem);
+            removeOwnerMetadata(context, profileItem);
         }
 
     }
@@ -161,34 +159,52 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
     }
 
     @Override
-    public ResearcherProfile claim(final Context context, final EPerson ePerson, final URI uri)
+    public ResearcherProfile claim(Context context, EPerson ePerson, URI uri)
             throws SQLException, AuthorizeException, SearchServiceException {
 
         Item profileItem = findResearcherProfileItemById(context, ePerson.getID());
         if (profileItem != null) {
-            ResearcherProfile profile = new ResearcherProfile(profileItem);
-            throw new ResourceConflictException("A profile is already linked to the provided User", profile);
+            throw new ResourceAlreadyExistsException("A profile is already linked to the provided User");
         }
 
         Item item = findItemByURI(context, uri)
             .orElseThrow(() -> new IllegalArgumentException("No item found by URI " + uri));
 
-        if (!item.isArchived() || item.isWithdrawn() || notClaimableEntityType(item)) {
-            throw new IllegalArgumentException("Provided uri does not represent a valid Item to be claimed");
+        if (!item.isArchived() || item.isWithdrawn()) {
+            throw new IllegalArgumentException(
+                "Only archived items can be claimed to create a researcher profile. Item ID: " + item.getID());
+        }
+
+        if (!hasProfileType(item)) {
+            throw new IllegalArgumentException("The provided item has not a profile type. Item ID: " + item.getID());
         }
 
         String existingOwner = itemService.getMetadataFirstValue(item, "dspace", "object", "owner", Item.ANY);
 
         if (StringUtils.isNotBlank(existingOwner)) {
-            throw new IllegalArgumentException("Item with provided uri has already an owner");
+            throw new IllegalArgumentException("Item with provided uri has already an owner - ID: " + existingOwner);
         }
 
         context.turnOffAuthorisationSystem();
         itemService.addMetadata(context, item, "dspace", "object", "owner", null,
                                 ePerson.getName(), ePerson.getID().toString(), CF_ACCEPTED);
-
         context.restoreAuthSystemState();
+
         return new ResearcherProfile(item);
+    }
+
+    @Override
+    public boolean hasProfileType(Item item) {
+        String profileType = getProfileType();
+        if (StringUtils.isBlank(profileType)) {
+            return false;
+        }
+        return profileType.equals(itemService.getEntityType(item));
+    }
+
+    @Override
+    public String getProfileType() {
+        return configurationService.getProperty("researcher-profile.entity-type", "Person");
     }
 
     private Optional<Item> findItemByURI(final Context context, final URI uri) throws SQLException {
@@ -197,11 +213,9 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
         return ofNullable(itemService.find(context, uuid));
     }
 
-    private boolean notClaimableEntityType(final Item item) {
-        String entityType = itemService.getEntityType(item);
-        return !contains(configurationService.getArrayProperty("claimable.entityType"), entityType);
-    }
-
+    /**
+     * Search for an profile item owned by an eperson with the given id.
+     */
     private Item findResearcherProfileItemById(Context context, UUID id) throws SQLException, AuthorizeException {
 
         String profileType = getProfileType();
@@ -218,6 +232,10 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
         return null;
     }
 
+    /**
+     * Returns a Profile collection based on a configuration or searching for a
+     * collection of researcher profile type.
+     */
     @SuppressWarnings("rawtypes")
     private Optional<Collection> findProfileCollection(Context context) throws SQLException, SearchServiceException {
         UUID uuid = UUIDUtils.fromString(configurationService.getProperty("researcher-profile.collection.uuid"));
@@ -246,6 +264,9 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
         return ofNullable((Collection) indexableObjects.get(0).getIndexedObject());
     }
 
+    /**
+     * Create a new profile item for the given ePerson in the provided collection.
+     */
     private Item createProfileItem(Context context, EPerson ePerson, Collection collection)
             throws AuthorizeException, SQLException {
 
@@ -259,20 +280,32 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
 
         item = installItemService.installItem(context, workspaceItem);
 
-        Group anonymous = groupService.findByName(context, ANONYMOUS);
-        authorizeService.removeGroupPolicies(context, item, anonymous);
+        if (isNewProfilePrivateByDefault()) {
+            Group anonymous = groupService.findByName(context, ANONYMOUS);
+            authorizeService.removeGroupPolicies(context, item, anonymous);
+        }
+
         authorizeService.addPolicy(context, item, READ, ePerson);
 
-        return item;
+        return reloadItem(context, item);
     }
 
     private boolean isHardDeleteEnabled() {
         return configurationService.getBooleanProperty("researcher-profile.hard-delete.enabled");
     }
 
-    private void removeDspaceObjectOwnerMetadata(Context context, Item profileItem) throws SQLException {
+    private boolean isNewProfilePrivateByDefault() {
+        return configurationService.getBooleanProperty("researcher-profile.set-new-profile-private");
+    }
+
+    private void removeOwnerMetadata(Context context, Item profileItem) throws SQLException {
         List<MetadataValue> metadata = itemService.getMetadata(profileItem, "dspace", "object", "owner", Item.ANY);
         itemService.removeMetadataValues(context, profileItem, metadata);
+    }
+
+    private Item reloadItem(Context context, Item item) throws SQLException {
+        context.uncacheEntity(item);
+        return context.reloadEntity(item);
     }
 
     private void deleteItem(Context context, Item profileItem) throws SQLException, AuthorizeException {
@@ -284,10 +317,6 @@ public class ResearcherProfileServiceImpl implements ResearcherProfileService {
         } finally {
             context.restoreAuthSystemState();
         }
-    }
-
-    private String getProfileType() {
-        return configurationService.getProperty("researcher-profile.type", "Person");
     }
 
 }
